+++ src/main/java/ru/sberhealth/rgs/Homework.java
@@ -1,0 +1,34 @@
+package ru.sberhealth.rgs;
+
+public class Homework {
+
+    public static void main(String[] args) {
+        // Создать метод, который принимает аргумент типа String имя пользователя, а также аргумент типа int возраст пользователя
+        // метод должен возвращать значение типа String, в котором упоминается имя пользователя и его возраст. Например
+        // "Привет, Анатолий! Рады сообщить, что тебе на сегодняшний день 24 года!". Вывести в консоль содержимое строки
+
+        // Создать метод, который вычисляет возраст пользователя по его году рождения. Метод должен возвращать число,
+        // которое соответствует тому, сколько на данный момент полных лет пользователю. Также вывести на экран это значение
+
+        // Создать метод, который из двух строк делает одну и возвращает новое значение в виде строки. Вывести результат в консоль
+
+        // Создать метод, который принимает в себя аргумент типа String, в котором должно быть число без дробной части, и возвращает его
+        // в виде значения типа int. Вывести результат в консоль
+
+        // Создать метод, который принимает в себя аргумент типа String, в котором должно быть число с дробной частью, отделенной точкой, и возвращает его
+        // в виде значения типа double. Вывести результат в консоль
+
+        // Создать метод, который принимает в себя аргумент типа String, в котором должно быть число с дробной частью, отделенной точкой, и возвращает его
+        // в виде округленного значения типа int. Вывести результат в консоль
+
+        // Используя java.lang.Math посчитать квадратный корень для числа 47 и вывести на экран
+
+        // Используя java.lang.Math вывести на экран случайное число типа double, которое будет в границах 0 < число < 1
+
+        // Используя java.lang.Math вывести на экран большее из двух чисел - 13 и 17
+
+        // Используя java.lang.Math вывести на экран меньшее из двух чисел - 4.5 и 3.8
+
+        // Используя java.lang.Math вывести на экран округленные числа 4.5, 5.6 и 2.3
+    }
+}
